Move Checkin Business Logic:
- Move checkin business logic to separate package and update API to call it to execution actions.
