+++ backend/app/checkin/__init__.py
@@ -1,0 +1,5 @@
+from .Checkin import CheckinLogic
+
+checkin_logic = CheckinLogic()
+
+__all__ = ["checkin_logic"]
